Assign project copyright to Raven Industries
There's a team of engineers at Raven Industries that wanted to work on
an open sourced, MIT (or BSD) licensed Rust CAN stack for their
innovation sprint project. We got approval from Raven's leadership and
legal teams for doing so, under the provision that Raven maintains
copyright of the resulting project.

However, those same engineers believe the project will be most
successful long-term if we work together with the _excellent_
Open-Agriculture community, which seems especially true as the very same
community has expressed significant interest in such a CAN stack
[1,2,3,4] over the last few years. This belief is founded on the
momentum we have observed from the Open-Agriculture community, as well
as a believe in the values of contributing to open-source software.

See also the discussion in [5] for more context.

[1] https://github.com/Open-Agriculture/AgIsoStack-plus-plus/discussions/215
[2] https://github.com/Open-Agriculture/AgIsoStack-plus-plus/discussions/279
[3] https://github.com/Thom-de-Jong/AgIsoStack-rs
[4] https://github.com/OpenIsobus/OpenIsobus
[5] https://github.com/Open-Agriculture/AgIsoStack-rs/discussions/1