Fix broken CI pipeline on main
The PR coverage report step uses a bot to add a comment on the PR
associated with this commit. But for it to do so, it requires finding
the PR number for the current branch. I found two ways to do so:

1. Change "on: [push]" to "on: [pull_request]" at the top-level
   workflow.
2. Use the "gh" CLI tool (which is installed by default in every default
   runs-on: distribution that GitHub provides) to find the PR number for
   the current branch

I picked option 2, because option 1 means that the workflow is only ever
run on the creation of a PR, and not on every push to an existing PR.

But I didn't think about the pipeline that will run on the main branch
after merging the PR, in which case there is no PR to submit a
code-coverage comment on.