Require the CAN frame type when creating a CanId
The most reliable way to know if a CAN ID is standard or extended is for
the driver to determine this (based on the IDE bit, which isn't
contained in the 29-bit ID). So we require the driver to determine the
frame type when creating the ID.

This was supposed to have been merged as a part of [1], but because of a
rebase-and-force-push snafu, probably as a result of poor Git hygiene on
my part, these changes were overwritten before the PR was merged.

[1] https://github.com/Open-Agriculture/AgIsoStack-rs/pull/4