Pin the CI/CD Rust toolchain to latest stable
A frequent source of CI failures (here, but mostly elsewhere) is not
pinning your compiler version. As new versions are released, new lints
and errors are added, making a project that used to build without
errors, now fail simply because there's a new release.

Now, we can, and SHOULD frequently upgrade our toolchain, but it should
still be pinned, so that we can do it on our terms.

Closes #36.